# app/internal/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateHallRequest:
    properties:
      branch_id:
        type: string
      name:
        type: string
      name_i18n:
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  model.HallResponse:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      name_i18n:
        type: string
      updated_at:
        type: string
    type: object
...
  model.UpdateHallRequest:
    properties:
      branch_id:
        type: string
      name:
        type: string
      name_i18n:
        type: string
    type: object
...
host: back.staging.maryai.yurtal.tech
info:
  contact: {}
  description: MaryAI API server with multi-language support (uz, ru, en)
  title: MaryAI API
  version: "1.0"
paths:
...
  /api/v1/halls:
    get:
      consumes:
      - application/json
      description: Retrieve all halls with pagination
      parameters:
      - description: 'Limit (default: 20)'
        in: query
        name: limit
        type: integer
      - description: 'Offset (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Halls found
          schema:
            items:
              $ref: '#/definitions/model.HallResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all halls
      tags:
      - halls
    post:
      consumes:
      - application/json
      description: Create a new hall with name, branch ID, and optional translation
        ID
      parameters:
      - description: Hall creation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateHallRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Hall created successfully
          schema:
            $ref: '#/definitions/model.HallResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new hall
      tags:
      - halls
  /api/v1/halls/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a hall by ID
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Hall deleted successfully
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Hall not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a hall
      tags:
      - halls
    get:
      consumes:
      - application/json
      description: Retrieve a specific hall by its ID
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Hall found
          schema:
            $ref: '#/definitions/model.HallResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Hall not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a hall by ID
      tags:
      - halls
    put:
      consumes:
      - application/json
      description: Update an existing hall
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      - description: Hall update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UpdateHallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Hall updated successfully
          schema:
            $ref: '#/definitions/model.HallResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Hall not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a hall
      tags:
      - halls
  /api/v1/halls/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restore a soft-deleted hall by ID
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Hall restored successfully
          schema:
            $ref: '#/definitions/model.HallResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Hall not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore a hall
      tags:
      - halls
  /api/v1/halls/branch/{branchId}:
    get:
      consumes:
      - application/json
      description: Retrieve all halls for a specific branch
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: string
      - description: 'Limit (default: 20)'
        in: query
        name: limit
        type: integer
      - description: 'Offset (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Halls found
          schema:
            items:
              $ref: '#/definitions/model.HallResponse'
            type: array
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get halls by branch ID
      tags:
      - halls
  /api/v1/halls/search:
    get:
      consumes:
      - application/json
      description: Search for halls by name with pagination
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: 'Limit (default: 20)'
        in: query
        name: limit
        type: integer
      - description: 'Offset (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Halls found
          schema:
            items:
              $ref: '#/definitions/model.HallResponse'
            type: array
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search halls
      tags:
      - halls
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"